fixup! Add UnionStatsRule
